+++ context/app/static/js/components/Detail/StatusIcon/style.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import LensIcon from '@material-ui/icons/Lens';
+import LensIcon from '@material-ui/icons/LensRounded';
 
 const ColoredStatusIcon = styled(LensIcon)`
   color: ${(props) => props.theme.palette[props.$iconColor].main};