+++ context/app/static/js/components/Detail/Summary/style.js
@@ -1,28 +1,9 @@
 import styled from 'styled-components';
 import Paper from '@material-ui/core/Paper';
 
-const FlexContainer = styled.div`
-  display: flex;
-`;
-
-const FlexCenterAlign = styled(FlexContainer)`
-  align-items: center;
-`;
-
-const FlexColumn = styled(FlexContainer)`
-  flex-direction: column;
-`;
-
-const FlexRight = styled.div`
+const FlexColumnRight = styled.div`
   display: flex;
   margin-left: auto;
-`;
-
-const FlexBottomRight = styled(FlexRight)`
-  align-items: flex-end;
-`;
-
-const FlexColumnRight = styled(FlexRight)`
   flex-direction: column;
   justify-content: space-evenly;
   white-space: nowrap;
@@ -34,4 +15,8 @@ const StyledPaper = styled(Paper)`
   padding: 30px 40px 30px 40px;
 `;
 
-export { FlexContainer, FlexCenterAlign, FlexColumn, FlexRight, FlexBottomRight, FlexColumnRight, StyledPaper };
+const StyledDiv = styled.div`
+  margin-bottom: ${(props) => props.theme.spacing(1)}px;
+`;
+
+export { FlexColumnRight, StyledPaper, StyledDiv };