+++ context/test-utils/setupTests.js
@@ -1,2 +1,3 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import '@testing-library/jest-dom/extend-expect';
+import 'whatwg-fetch'; // polyfill fetch